+++ src/main/java/jumper/filter/RequestFilter.java
@@ -107,12 +107,6 @@ public class RequestFilter extends AbstractGatewayFilterFactory<RequestFilter.Co
 
                 // no failover
                 else {
-                  // checking to prevent later nullPointer on inconsistent state from Kong
-                  if (!request.getHeaders().containsKey(Constants.HEADER_REMOTE_API_URL)) {
-                    throw new RuntimeException(
-                        "missing mandatory header " + Constants.HEADER_REMOTE_API_URL);
-                  }
-
                   // Prepare and extract JumperConfigValues
                   jumperConfig = JumperConfig.parseAndFillJumperConfigFrom(request);
                   log.debug("JumperConfig decoded: {}", jumperConfig);
@@ -129,9 +123,15 @@ public class RequestFilter extends AbstractGatewayFilterFactory<RequestFilter.Co
                       .put(Constants.HEADER_JUMPER_CONFIG, JumperConfig.toBase64(jumperConfig));
                 }
 
-                // write audit log if needed
-                if (jumperConfig.getAuditLog()) {
+                if (jumperConfig.getSecondaryFailover()) {
+                  // write audit log if needed
                   AuditLogService.writeFailoverAuditLog(jumperConfig);
+
+                  // pass headers from config to provider
+                  HeaderUtil.addHeader(
+                      exchange, Constants.HEADER_REALM, jumperConfig.getRealmName());
+                  HeaderUtil.addHeader(
+                      exchange, Constants.HEADER_ENVIRONMENT, jumperConfig.getEnvName());
                 }
 
                 // handle request
@@ -466,7 +466,7 @@ public class RequestFilter extends AbstractGatewayFilterFactory<RequestFilter.Co
     for (JumperConfig jc : jumperConfigList) {
       // secondary route, failover in place => audit logs
       if (StringUtils.isEmpty(jc.getTargetZoneName())) {
-        jc.setAuditLog(true);
+        jc.setSecondaryFailover(true);
         return jc;
       }
       // targetZoneName present, check it against force skip header and zones state