+++ src/test/java/jumper/util/TokenUtil.java
@@ -100,6 +100,17 @@ public class TokenUtil {
     };
   }
 
+  public static Consumer<HttpHeaders> getRealRouteHeadersWithoutRemoteApiUrl(String authorization) {
+    return httpHeaders -> {
+      if (authorization != null) httpHeaders.setBearerAuth(authorization);
+      httpHeaders.set(Constants.HEADER_API_BASE_PATH, BASE_PATH);
+      httpHeaders.set(Constants.HEADER_ENVIRONMENT, ENVIRONMENT);
+      httpHeaders.set(Constants.HEADER_REALM, REALM);
+      httpHeaders.set(Constants.HEADER_ACCESS_TOKEN_FORWARDING, "false");
+      httpHeaders.set(Constants.HEADER_JUMPER_CONFIG, "e30=");
+    };
+  }
+
   public static Consumer<HttpHeaders> getEmptyHeaders() {
     return httpHeaders -> {};
   }